refactor: 코드 품질 개선 및 경고 제거
Rust clippy 경고 수정:
- 자동 수정 가능한 9개 경고 자동 적용
- dead_code 경고: get_current_path에 #[allow(dead_code)] 추가
- enum_variant_names 경고: FolderChangeEvent에 #[allow] 추가

하드코딩된 값을 상수로 추출:
- FOLDER_WATCH_RESUME_DELAY 상수 추가 (600ms)
- ThumbnailPanel, ImageViewerPanel에서 매직 넘버 제거
- 폴더 감시 재개 딜레이를 상수로 통일

코드 일관성 향상:
- 주석 개선 (500ms -> 디바운스 시간으로 일반화)
- 두 패널 간 동일한 패턴 적용

fix: thumbnail.rs 불필요한 괄호 제거

+++ src-tauri/src/thumbnail.rs
@@ -589,8 +589,8 @@ fn extract_webp_dimensions(webp_data: &[u8]) -> Option<(u32, u32)> {
                 return None;
             }
             let bits = u32::from_le_bytes([webp_data[21], webp_data[22], webp_data[23], webp_data[24]]);
-            let width = ((bits & 0x3FFF) + 1) as u32;
-            let height = (((bits >> 14) & 0x3FFF) + 1) as u32;
+            let width = (bits & 0x3FFF) + 1;
+            let height = ((bits >> 14) & 0x3FFF) + 1;
             Some((width, height))
         }
         b"VP8X" => {
